Some layout fixes, caching for hosts

+++ app/src/main/java/com/example/david/gigfinder/HostProfileFragment.java
@@ -67,7 +67,7 @@ public class HostProfileFragment extends Fragment {
     private static final int RESULT_EDIT_PROFILE = 1;
     private static final int RESULT_PICK_IMAGE = 2;
 
-    SharedPreferences sharedPreferences;
+    private SharedPreferences sharedPreferences;
 
     private int userID;
     private Button testDeleteBtn;
@@ -300,17 +300,11 @@ public class HostProfileFragment extends Fragment {
             JSONArray jsonArray = new JSONArray(jsonString);
             JSONObject userProfile = jsonArray.getJSONObject(0);
 
-            if(!idToken.equals("offline")) {
-                if(isNetworkAvailable()) {
+            GetProfilePicture getProfilePicture = new GetProfilePicture();
+            getProfilePicture.execute(userProfile.getInt("profilePictureId") + "");
 
-                    GetProfilePicture getProfilePicture = new GetProfilePicture();
-                    getProfilePicture.execute(userProfile.getInt("profilePictureId") + "");
-
-                    GetReview getReview = new GetReview();
-                    getReview.execute();
-
-                }
-            }
+            GetReview getReview = new GetReview();
+            getReview.execute();
 
             final String name = userProfile.getString("name");
             nameText.setText(name);
@@ -702,7 +696,7 @@ public class HostProfileFragment extends Fragment {
     /**
      * Gets the profile picture from the server and GUI update
      */
-    class GetProfilePicture extends AsyncTask<String, Void, String> {
+    private class GetProfilePicture extends AsyncTask<String, Void, String> {
 
         @Override
         protected String doInBackground(String... params) {
@@ -712,8 +706,11 @@ public class HostProfileFragment extends Fragment {
 
                 urlConnection.setRequestProperty("Authorization", idToken);
                 urlConnection.setRequestMethod("GET");
-                urlConnection.setUseCaches(true);
-                urlConnection.addRequestProperty("Cache-Control", "max-stale="+getString(R.string.max_stale_online));
+                if(isNetworkAvailable()) {
+                    urlConnection.addRequestProperty("Cache-Control", "max-stale=" + getString(R.string.max_stale_online));
+                }else{
+                    urlConnection.addRequestProperty("Cache-Control", "max-stale=" + getString(R.string.max_stale_offline));
+                }
 
                 BufferedReader in = new BufferedReader(new InputStreamReader(urlConnection.getInputStream()));
                 String inputLine;
@@ -742,8 +739,7 @@ public class HostProfileFragment extends Fragment {
         }
     }
 
-
-    class PutProfilePicture extends AsyncTask<String, Void, String> {
+    private class PutProfilePicture extends AsyncTask<String, Void, String> {
 
         @Override
         protected String doInBackground(String... params) {
@@ -813,7 +809,7 @@ public class HostProfileFragment extends Fragment {
     /**
      *
      */
-    class DeleteUser extends AsyncTask<String, Void, String> {
+    private class DeleteUser extends AsyncTask<String, Void, String> {
 
         @Override
         protected String doInBackground(String... params) {
@@ -847,27 +843,29 @@ public class HostProfileFragment extends Fragment {
 
         @Override
         protected void onPostExecute(String result) {
-            SharedPreferences.Editor editor = sharedPreferences.edit();
-            editor.clear();
-            editor.apply();
-            Log.d(TAG, "DELETE HOST: " + result);
-            Intent intent = new Intent(getActivity(), LoginActivity.class);
-            intent.putExtra("SignOut", true);
-            startActivity(intent);
+            if (result != null) {
+                SharedPreferences.Editor editor = sharedPreferences.edit();
+                editor.clear();
+                editor.apply();
+                Log.d(TAG, "DELETE HOST: " + result);
+                Intent intent = new Intent(getActivity(), LoginActivity.class);
+                intent.putExtra("SignOut", true);
+                startActivity(intent);
 
-            try {
-                HttpResponseCache cache = HttpResponseCache.getInstalled();
-                cache.delete();
-            } catch (IOException e) {
-                e.printStackTrace();
+                try {
+                    HttpResponseCache cache = HttpResponseCache.getInstalled();
+                    cache.delete();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+
+                Toast.makeText(getActivity().getApplicationContext(), "Profil Gelöscht", Toast.LENGTH_SHORT).show();
+                getActivity().finish();
             }
-
-            Toast.makeText(getActivity().getApplicationContext(),"Profil Gelöscht",Toast.LENGTH_SHORT).show();
-            getActivity().finish();
         }
     }
 
-    class GetReview extends AsyncTask<String, Void, String> {
+    private class GetReview extends AsyncTask<String, Void, String> {
 
         @Override
         protected String doInBackground(String... params) {
@@ -877,6 +875,9 @@ public class HostProfileFragment extends Fragment {
 
                 urlConnection.setRequestProperty("Authorization", idToken);
                 urlConnection.setRequestMethod("GET");
+                if(!isNetworkAvailable()){
+                    urlConnection.addRequestProperty("Cache-Control", "max-stale=" + getString(R.string.max_stale_offline));
+                }
 
                 BufferedReader in = new BufferedReader(new InputStreamReader(urlConnection.getInputStream()));
                 String inputLine;
